moved suggestion generation from viewmodel into new class
