+++ NetPrintsEditor/Reflection/ISuggestionGenerator.cs
@@ -1,0 +1,10 @@
+﻿using NetPrints.Base;
+using System.Collections.Generic;
+
+namespace NetPrintsEditor.Reflection
+{
+    public interface ISuggestionGenerator
+    {
+        IEnumerable<(string, object)> GetSuggestions(INodeGraph graph, INodePin pin = null);
+    }
+}
